Convert SideNav file to TS
