+++ src/components/DocsNav.tsx
@@ -5,7 +5,8 @@ import Link from "./Link"
 import Emoji from "./Emoji"
 import Translation from "./Translation"
 
-import docLinks, { DeveloperDocsLink } from "../data/developer-docs-links.yaml"
+import docLinks from "../data/developer-docs-links.yaml"
+import { DeveloperDocsLink } from "../types"
 import { TranslationKey } from "../utils/translations"
 
 const Container = styled.div`
