+++ src/components/SideNav.tsx
@@ -10,6 +10,7 @@ import { dropdownIconContainerVariant } from "./SharedStyledComponents"
 
 import docLinks from "../data/developer-docs-links.yaml"
 import { translateMessageId } from "../utils/translations"
+import { DeveloperDocsLink } from "../types"
 
 const IconContainer = styled(motion.div)`
   cursor: pointer;
@@ -79,9 +80,14 @@ const SideNavGroup = styled.div`
 
 const NavItem = styled.div``
 
-const NavLink = ({ item, path }) => {
+export interface IPropsNavLink {
+  item: DeveloperDocsLink
+  path: string
+}
+
+const NavLink: React.FC<IPropsNavLink> = ({ item, path }) => {
   const isLinkInPath = path.includes(item.to) || path.includes(item.path)
-  const [isOpen, setIsOpen] = useState(isLinkInPath)
+  const [isOpen, setIsOpen] = useState<boolean>(isLinkInPath)
 
   useEffect(() => {
     // Only set on items that contain a link
@@ -138,11 +144,15 @@ const NavLink = ({ item, path }) => {
   )
 }
 
+export interface IProps {
+  path: string
+}
+
 // TODO set tree state based on if current path is a child
 // of the given parent. Currently all `path` items default to open
 // and they only collapse when clicked on.
 // e.g. solution: https://github.com/hasura/gatsby-gitbook-starter/blob/5c165af40e48fc55eb06b45b95c84eb64b17ed32/src/components/sidebar/tree.js
-const SideNav = ({ path }) => {
+const SideNav: React.FC<IProps> = ({ path }) => {
   const intl = useIntl()
 
   return (
