+++ src/types.ts
@@ -1,5 +1,6 @@
 import type { Messages } from "./interfaces"
 import type { Lang } from "./utils/languages"
+import { TranslationKey } from "./utils/translations"
 
 export type Intl = {
   language: Lang
@@ -20,6 +21,14 @@ export type Context = {
   isOutdated: boolean
   isLegal?: boolean
   isContentEnglish?: boolean
+}
+
+export interface DeveloperDocsLink {
+  id: TranslationKey
+  to: string
+  path: string
+  description: TranslationKey
+  items: Array<DeveloperDocsLink>
 }
 
 export type Direction = "rtl" | "ltr" | "auto"
